add apps/afc subcommand
